Sort events by start date, earliest to latest
To make this sorting easier, we’ve removed the separate date and time
fields from the event post_meta, and instead combined them into a single
date_time field, one for start, one for end. This means we can no longer
have an event with an unspecified start *time* (but a specified start
*date*) but it keeps everything else simpler, so worth it, I think.

Note, also, that if there are lots of past events in the database,
they’ll be displayed first in the list, which isn’t ideal. We really
need to filter out past events and provide a way to view them separate
from upcoming events.